requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
pypdf>=3.17.0
tqdm>=4.66.0
# AI Text Generation (GPT-2)
transformers>=4.30.0
torch>=2.0.0
# PDF Document Generation
reportlab>=4.0.0
